Pull adapter to a separate class, add identifier and number of packages, minor UI cleanup
