+++ purchases-sample/src/main/res/values/styles.xml
@@ -6,6 +6,11 @@
         <item name="colorPrimary">@color/colorPrimary</item>
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
+        <item name="materialCardViewStyle">@style/MaterialCardStyle</item>
     </style>
 
+    <style name="MaterialCardStyle" parent="Widget.MaterialComponents.CardView">
+        <item name="cardCornerRadius">@dimen/padding_tiny</item>
+        <item name="cardElevation">@dimen/padding_tiny</item>
+    </style>
 </resources>
